Select and apply expression to elements of table like lists.

Each list element is attached to both extractor and applicator environment so its own named elements are available as variables prefixed with "." (dot character).

inlist(l, extractor, applicator, fallback)

Arguments

l

list (or NULL)

extractor

Predicate expression to filter list elements before applying evaluated in the environment of list's element

applicator

Expression applied to each element of the list evaluated in the element's environment

fallback

Optional. If provided use this value as a fall back in case some variables (prefixed with dot) are not fount in the list's element environment. Otherwise (the default) those elements will be ignored and not included to results

Value

List of filtered with extractor elements with values returned by applicator. Unbound expressions are omitted (if fallback is not provided) so list might be shorter.

Details

. - current list's element

.. - (filtered) list (in extractor environment it is the same as initial list)

.n - (filtered) list element's index

.N - (filtered) list's length

... - initial list (l)

..n - initial list element's index

..N - initial list's length

._(x, fb) - function that returns the fallback fb if some of the variables in x are unbound (i.e., due to non existing list elements)